Generate moisture with diamond square and remove unnecessary code

+++ mapGenerator/src/main/java/mapgenerator/domain/Map.java
@@ -8,7 +8,7 @@ public class Map {
     
     private double[][] heightMap;
     private boolean[][] water;
-    private int[][] moisture;
+    private double[][] moisture;
     
     /**
      * Sets a height map.
@@ -42,11 +42,11 @@ public class Map {
         return water;
     }
     
-    public void setMoisture(int[][] moisture) {
+    public void setMoisture(double[][] moisture) {
         this.moisture = moisture;
     }
     
-    public int[][] getMoisture() {
+    public double[][] getMoisture() {
         return moisture;
     }
